chore: moved session and document collection to shared
