+++ packages/shared/src/payload.ts
@@ -1,10 +1,19 @@
 /* -----------------------------------
  *
+ * Variables
+ *
+ * -------------------------------- */
+
+const clientKey = 'clientId';
+const sessionKey = 'sessionId';
+
+/* -----------------------------------
+ *
  * Document
  *
  * -------------------------------- */
 
-function getDocumentMeta() {
+function getDocument() {
   const title = document.title;
   const hostname = document.location.hostname;
   const origin = document.location.origin;
@@ -17,8 +26,46 @@ function getDocumentMeta() {
 
 /* -----------------------------------
  *
+ * ClientId
+ *
+ * -------------------------------- */
+
+function getClientId(key = clientKey) {
+  const clientId = Math.random().toString(36);
+  const storedValue = localStorage.getItem(key);
+
+  if (!storedValue) {
+    localStorage.setItem(key, clientId);
+
+    return clientId;
+  }
+
+  return storedValue;
+}
+
+/* -----------------------------------
+ *
+ * SessionId
+ *
+ * -------------------------------- */
+
+function getSessionId(key = sessionKey) {
+  const sessionId = `${Math.floor(Math.random() * 1000000000) + 1}`;
+  const storedValue = sessionStorage.getItem(key);
+
+  if (!storedValue) {
+    sessionStorage.setItem(key, sessionId);
+
+    return sessionId;
+  }
+
+  return storedValue;
+}
+
+/* -----------------------------------
+ *
  * Export
  *
  * -------------------------------- */
 
-export { getDocumentMeta };
+export { getDocument, getClientId, getSessionId };
